% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatureset.R
\name{addFeatureInformation}
\alias{addFeatureInformation}
\alias{listFeatureInformation}
\alias{removeFeatureInformation}
\title{Add, list and remove feature information}
\usage{
addFeatureInformation(task, which, features)

listFeatureInformation(task)

removeFeatureInformation(task)
}
\arguments{
\item{task}{[\code{\link[mlr]{Task}}]\cr
The task.}

\item{which}{[character(1)]\cr
Kind of feature information can be \code{timestamp}, \code{categ.featureset}, \code{numeric.featureset}.}

\item{features}{[character()]\cr
Features to be added to the kind of feature information.}
}
\value{
[\code{\link[mlr]{Task}}]
}
\description{
Add additional information about features to a task. \code{\link{autoxgboost}} will
do feature engineering based on this information.
}
\section{Functions}{
\itemize{
\item \code{listFeatureInformation}: list feature information

\item \code{removeFeatureInformation}: delete feature information
}}

\examples{
\donttest{
iris.task = makeClassifTask(data = iris, target = "Species")
iris.task = addFeatureInformation(iris.task, "numeric.featuresets",
  c("Sepal.Length", "Sepal.Width"))
listFeatureInformation(iris.task)
removeFeatureInformation(iris.task)
}
}
