% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoxgbResult.R
\name{AutoxgbResult}
\alias{AutoxgbResult}
\title{Result of a autoxgboost call.}
\description{
\itemize{
  \item{optim.result [\code{\link[mlrMBO]{MBOResult}}]}{Optimization result object. See: \code{\link[mlrMBO]{MBOResult}}}
  \item{optim.result [\code{\link[mlr]{MBOResult}}]}{Optimization result object. See: \code{\link[mlrMBO]{MBOResult}}}

}
}

