% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactFeatureWrapper.R
\name{makeImpactFeaturesWrapper}
\alias{makeImpactFeaturesWrapper}
\title{Fuse learner with impact feature creator.}
\usage{
makeImpactFeaturesWrapper(learner, cols = NULL, fun = NULL,
  slope.param = 100L, trust.param = 100L)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{cols}{[\code{character}]\cr
Columns to create impact features for. Default is to use all columns.}

\item{fun}{[\code{function}]\cr
Function to apply to the response for each factor level in order to obtain the 
conditional probabilities for each factor level, should always return a numeric vector.
If \code{NULL}, \code{mean} is used for regression, \code{table(x)[1] / length(x)} for binary classification,
and \code{table(x) / length(x)} for multiclass classification.}

\item{slope.param}{[\code{integer}]\cr
Controls the rate of transition rate \eqn{\lambda}. Default is \code{100L}.}

\item{trust.param}{[\code{integer}]\cr
Determines half of the minimal sample size for which we completely "trust" the conditional 
probability of a factor level. Default is \code{100L}}
}
\value{
[\code{\link{Learner}}].
}
\description{
Fuses a base learner with the impact feature creator (see \code{\link{createImpactFeatures}}).
Returns a learner which can be used like any other learner.
}
