% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegerFeatureWrapper.R
\name{makeIntegerFeaturesWrapper}
\alias{makeIntegerFeaturesWrapper}
\title{Fuse learner with integer feature creator.}
\usage{
makeIntegerFeaturesWrapper(learner, cols = NULL)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{cols}{[\code{character}]\cr
Columns to create integer features for. Default is to use all columns.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Changes all categorical features to integer values.
}
