% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoxgboost.R
\name{autoxgboost}
\alias{autoxgboost}
\title{Fit and optimize a xgboost model.}
\usage{
autoxgboost(task, measure, control, par.set = autoxgbparset)
}
\arguments{
\item{task}{[\code{\link[mlr]{Task}}]\cr
The task.}

\item{measure}{[\code{\link[mlr]{Measure}}]\cr
Performance measure.}

\item{control}{[\code{\link[mlrMBO]{MBOControl}}]\cr
Control object for mbo.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Parameter set.}
}
\description{
An xgboost model is optimized based on a measure (see [\code{\link[mlr]{Measure}}]).
The bounds of parameter in which the model is optimized, is defined by \code{autoxgbparset}.
For the optimization itself bayesian optimization with \pkg{mlrMBO} is used.
}

