% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoxgboost.R
\name{autoxgboost}
\alias{autoxgboost}
\title{Fit and optimize a xgboost model.}
\usage{
autoxgboost(task, measure = NULL, control = NULL, par.set = NULL,
  max.nrounds = 10^6, early.stopping.rounds = 10L,
  early.stopping.fraction = 4/5, build.final.model = TRUE,
  design.size = 15L, initial.subsample.range = c(0.5, 0.55))
}
\arguments{
\item{task}{[\code{\link[mlr]{Task}}]\cr
The task.}

\item{measure}{[\code{\link[mlr]{Measure}}]\cr
Performance measure. If \code{NULL} \code{\link[mlr]{getDefaultMeasure}} is used.}

\item{control}{[\code{\link[mlrMBO]{MBOControl}}]\cr
Control object for mbo. Specifies runtime behaviour.
Default is to run for 80 iterations or 1 hour, whatever happens first.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Parameter set. Default is \code{\link{autoxgbparset}}.}

\item{max.nrounds}{[\code{integer(1)}]\cr
Maximum number of allowed iterations. Default is \code{10^6}.}

\item{early.stopping.rounds}{[\code{integer(1L}]\cr
After how many iterations without an improvement in the OOB error should be stopped?
Default is 10.}

\item{early.stopping.fraction}{[\code{numeric(1)}]\cr
What fraction of the data should be used for early stopping (i.e. as a validation set).
Default is \code{4/5}.}

\item{build.final.model}{[\code{logical(1)}]\cr
Should the best found model be fitted on the complete dataset?
Default is \code{FALSE}.}

\item{design.size}{[\code{integer(1)}]\cr
Size of the initial design. Default is \code{15L}.}

\item{initial.subsample.range}{[\code{numeric(2)} | \code{NULL}]\cr
From what range should the subsampling parameter in the initial design be sampled?
It is useful to restrict to fraction to be rather small to speed up the calcuation of the initial design.
If \code{NULL} the full range defined in \code{par.set} is used.
Default is \code{c(0.5, 0.55)}.}
}
\value{
\code{\link{AutoxgbResult}}
}
\description{
An xgboost model is optimized based on a measure (see [\code{\link[mlr]{Measure}}]).
The bounds of the parameter in which the model is optimized, are defined by \code{\link{autoxgbparset}}.
For the optimization itself bayesian optimization with \pkg{mlrMBO} is used.
Without any specification of the control object, the optimizer runs for for 80 iterations or 1 hour, whatever happens first.
Both the parameter set and the control object can be set by the user.
}

