% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoxgbResult.R
\name{AutoxgbResult}
\alias{AutoxgbResult}
\title{Result of a autoxgboost call.}
\description{
\itemize{
  \item{optim.result [\code{\link[mlrMBO]{MBOSingleObjResult}}]}{Optimization result object. See: \code{\link[mlrMBO]{MBOSingleObjResult}}}
  \item{final.learner [\code{\link[mlr]{Learner}}]}{Xgboost learner with best found hyper paramater configuration.}
  \item{final.model [\code{\link[mlr]{WrappedModel}} | \code{NULL}]}{If \code{build.final.model=TRUE} in \code{\link{autoxgboost}} a \pkg{mlr} model build by the full dataset and \code{final.learner}.}
  \item{measure [\code{\link[mlr]{Measure}}]}{Measure used for optimization.}
}
}
