% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactFeatureWrapper.R
\name{makeImpactFeaturesWrapper}
\alias{makeImpactFeaturesWrapper}
\title{Fuse learner with impact feature creator.}
\usage{
makeImpactFeaturesWrapper(learner, cols = NULL, fun = NULL)
}
\arguments{
\item{learner}{[\code{\link{Learner}} | \code{character(1)}]\cr
The learner.
If you pass a string the learner will be created via \code{\link{makeLearner}}.}

\item{cols}{[\code{character}]\cr
Columns to create impact features for. Default is to use all columns.}

\item{fun}{[\code{function}]\cr
Function to apply on the response for each factor level, should always return one numeric value.}
}
\value{
[\code{\link{Learner}}].
}
\description{
Fuses a base learner with the impact feature creator (see \code{\link{createImpactFeatures}}).
Returns a learner which can be used like any other learner.
}
